+++ data/astyx/ImageSets/test.txt
@@ -1,0 +1,546 @@
+000000
+000001
+000002
+000003
+000004
+000005
+000006
+000007
+000008
+000009
+000010
+000011
+000012
+000013
+000014
+000015
+000016
+000017
+000018
+000019
+000020
+000021
+000022
+000023
+000024
+000025
+000026
+000027
+000028
+000029
+000030
+000031
+000032
+000033
+000034
+000035
+000036
+000037
+000038
+000039
+000040
+000041
+000042
+000043
+000044
+000045
+000046
+000047
+000048
+000049
+000050
+000051
+000052
+000053
+000054
+000055
+000056
+000057
+000058
+000059
+000060
+000061
+000062
+000063
+000064
+000065
+000066
+000067
+000068
+000069
+000070
+000071
+000072
+000073
+000074
+000075
+000076
+000077
+000078
+000079
+000080
+000081
+000082
+000083
+000084
+000085
+000086
+000087
+000088
+000089
+000090
+000091
+000092
+000093
+000094
+000095
+000096
+000097
+000098
+000099
+000100
+000101
+000102
+000103
+000104
+000105
+000106
+000107
+000108
+000109
+000110
+000111
+000112
+000113
+000114
+000115
+000116
+000117
+000118
+000119
+000120
+000121
+000122
+000123
+000124
+000125
+000126
+000127
+000128
+000129
+000130
+000131
+000132
+000133
+000134
+000135
+000136
+000137
+000138
+000139
+000140
+000141
+000142
+000143
+000144
+000145
+000146
+000147
+000148
+000149
+000150
+000151
+000152
+000153
+000154
+000155
+000156
+000157
+000158
+000159
+000160
+000161
+000162
+000163
+000164
+000165
+000166
+000167
+000168
+000169
+000170
+000171
+000172
+000173
+000174
+000175
+000176
+000177
+000178
+000179
+000180
+000181
+000182
+000183
+000184
+000185
+000186
+000187
+000188
+000189
+000190
+000191
+000192
+000193
+000194
+000195
+000196
+000197
+000198
+000199
+000200
+000201
+000202
+000203
+000204
+000205
+000206
+000207
+000208
+000209
+000210
+000211
+000212
+000213
+000214
+000215
+000216
+000217
+000218
+000219
+000220
+000221
+000222
+000223
+000224
+000225
+000226
+000227
+000228
+000229
+000230
+000231
+000232
+000233
+000234
+000235
+000236
+000237
+000238
+000239
+000240
+000241
+000242
+000243
+000244
+000245
+000246
+000247
+000248
+000249
+000250
+000251
+000252
+000253
+000254
+000255
+000256
+000257
+000258
+000259
+000260
+000261
+000262
+000263
+000264
+000265
+000266
+000267
+000268
+000269
+000270
+000271
+000272
+000273
+000274
+000275
+000276
+000277
+000278
+000279
+000280
+000281
+000282
+000283
+000284
+000285
+000286
+000287
+000288
+000289
+000290
+000291
+000292
+000293
+000294
+000295
+000296
+000297
+000298
+000299
+000300
+000301
+000302
+000303
+000304
+000305
+000306
+000307
+000308
+000309
+000310
+000311
+000312
+000313
+000314
+000315
+000316
+000317
+000318
+000319
+000320
+000321
+000322
+000323
+000324
+000325
+000326
+000327
+000328
+000329
+000330
+000331
+000332
+000333
+000334
+000335
+000336
+000337
+000338
+000339
+000340
+000341
+000342
+000343
+000344
+000345
+000346
+000347
+000348
+000349
+000350
+000351
+000352
+000353
+000354
+000355
+000356
+000357
+000358
+000359
+000360
+000361
+000362
+000363
+000364
+000365
+000366
+000367
+000368
+000369
+000370
+000371
+000372
+000373
+000374
+000375
+000376
+000377
+000378
+000379
+000380
+000381
+000382
+000383
+000384
+000385
+000386
+000387
+000388
+000389
+000390
+000391
+000392
+000393
+000394
+000395
+000396
+000397
+000398
+000399
+000400
+000401
+000402
+000403
+000404
+000405
+000406
+000407
+000408
+000409
+000410
+000411
+000412
+000413
+000414
+000415
+000416
+000417
+000418
+000419
+000420
+000421
+000422
+000423
+000424
+000425
+000426
+000427
+000428
+000429
+000430
+000431
+000432
+000433
+000434
+000435
+000436
+000437
+000438
+000439
+000440
+000441
+000442
+000443
+000444
+000445
+000446
+000447
+000448
+000449
+000450
+000451
+000452
+000453
+000454
+000455
+000456
+000457
+000458
+000459
+000460
+000461
+000462
+000463
+000464
+000465
+000466
+000467
+000468
+000469
+000470
+000471
+000472
+000473
+000474
+000475
+000476
+000477
+000478
+000479
+000480
+000481
+000482
+000483
+000484
+000485
+000486
+000487
+000488
+000489
+000490
+000491
+000492
+000493
+000494
+000495
+000496
+000497
+000498
+000499
+000500
+000501
+000502
+000503
+000504
+000505
+000506
+000507
+000508
+000509
+000510
+000511
+000512
+000513
+000514
+000515
+000516
+000517
+000518
+000519
+000520
+000521
+000522
+000523
+000524
+000525
+000526
+000527
+000528
+000529
+000530
+000531
+000532
+000533
+000534
+000535
+000536
+000537
+000538
+000539
+000540
+000541
+000542
+000543
+000544
+000545
