+++ data/astyx/ImageSets/train.txt
@@ -1,0 +1,269 @@
+000000
+000003
+000007
+000009
+000010
+000011
+000012
+000013
+000014
+000016
+000017
+000018
+000022
+000026
+000029
+000030
+000032
+000034
+000036
+000038
+000041
+000043
+000044
+000045
+000046
+000049
+000051
+000054
+000055
+000056
+000057
+000060
+000064
+000067
+000068
+000069
+000070
+000071
+000072
+000073
+000074
+000075
+000079
+000080
+000082
+000083
+000084
+000085
+000086
+000087
+000088
+000091
+000092
+000095
+000096
+000097
+000099
+000100
+000101
+000103
+000105
+000109
+000110
+000111
+000112
+000113
+000114
+000115
+000119
+000120
+000121
+000123
+000125
+000127
+000129
+000130
+000131
+000133
+000136
+000138
+000141
+000142
+000144
+000145
+000146
+000148
+000149
+000150
+000154
+000155
+000157
+000158
+000160
+000162
+000163
+000164
+000165
+000166
+000171
+000172
+000176
+000177
+000178
+000179
+000180
+000184
+000185
+000189
+000193
+000198
+000200
+000202
+000205
+000206
+000208
+000209
+000210
+000214
+000215
+000217
+000219
+000220
+000221
+000222
+000225
+000227
+000228
+000232
+000233
+000238
+000240
+000241
+000243
+000244
+000245
+000253
+000254
+000255
+000256
+000257
+000258
+000259
+000261
+000264
+000267
+000271
+000274
+000275
+000276
+000277
+000280
+000282
+000285
+000286
+000287
+000288
+000292
+000294
+000295
+000296
+000298
+000299
+000300
+000303
+000304
+000306
+000310
+000313
+000316
+000317
+000318
+000322
+000325
+000326
+000330
+000331
+000334
+000337
+000338
+000339
+000342
+000344
+000348
+000349
+000353
+000358
+000363
+000364
+000367
+000368
+000371
+000374
+000375
+000380
+000384
+000387
+000389
+000390
+000400
+000405
+000406
+000410
+000411
+000412
+000416
+000417
+000418
+000421
+000423
+000424
+000425
+000426
+000431
+000432
+000433
+000434
+000435
+000438
+000439
+000441
+000442
+000444
+000445
+000447
+000449
+000456
+000458
+000460
+000461
+000462
+000464
+000465
+000466
+000467
+000470
+000471
+000474
+000482
+000483
+000484
+000487
+000488
+000490
+000497
+000500
+000501
+000502
+000505
+000507
+000511
+000513
+000514
+000516
+000518
+000520
+000522
+000523
+000525
+000526
+000529
+000531
+000532
+000534
+000535
+000537
+000538
+000539
+000540
+000544
