+++ data/astyx/ImageSets/val.txt
@@ -1,0 +1,277 @@
+000001
+000002
+000004
+000005
+000006
+000008
+000015
+000019
+000020
+000021
+000023
+000024
+000025
+000027
+000028
+000031
+000033
+000035
+000037
+000039
+000040
+000042
+000047
+000048
+000050
+000052
+000053
+000058
+000059
+000061
+000062
+000063
+000065
+000066
+000076
+000077
+000078
+000081
+000089
+000090
+000093
+000094
+000098
+000102
+000104
+000106
+000107
+000108
+000116
+000117
+000118
+000122
+000124
+000126
+000128
+000132
+000134
+000135
+000137
+000139
+000140
+000143
+000147
+000151
+000152
+000153
+000156
+000159
+000161
+000167
+000168
+000169
+000170
+000173
+000174
+000175
+000181
+000182
+000183
+000186
+000187
+000188
+000190
+000191
+000192
+000194
+000195
+000196
+000197
+000199
+000201
+000203
+000204
+000207
+000211
+000212
+000213
+000216
+000218
+000223
+000224
+000226
+000229
+000230
+000231
+000234
+000235
+000236
+000237
+000239
+000242
+000246
+000247
+000248
+000249
+000250
+000251
+000252
+000260
+000262
+000263
+000265
+000266
+000268
+000269
+000270
+000272
+000273
+000278
+000279
+000281
+000283
+000284
+000289
+000290
+000291
+000293
+000297
+000301
+000302
+000305
+000307
+000308
+000309
+000311
+000312
+000314
+000315
+000319
+000320
+000321
+000323
+000324
+000327
+000328
+000329
+000332
+000333
+000335
+000336
+000340
+000341
+000343
+000345
+000346
+000347
+000350
+000351
+000352
+000354
+000355
+000356
+000357
+000359
+000360
+000361
+000362
+000365
+000366
+000369
+000370
+000372
+000373
+000376
+000377
+000378
+000379
+000381
+000382
+000383
+000385
+000386
+000388
+000391
+000392
+000393
+000394
+000395
+000396
+000397
+000398
+000399
+000401
+000402
+000403
+000404
+000407
+000408
+000409
+000413
+000414
+000415
+000419
+000420
+000422
+000427
+000428
+000429
+000430
+000436
+000437
+000440
+000443
+000446
+000448
+000450
+000451
+000452
+000453
+000454
+000455
+000457
+000459
+000463
+000468
+000469
+000472
+000473
+000475
+000476
+000477
+000478
+000479
+000480
+000481
+000485
+000486
+000489
+000491
+000492
+000493
+000494
+000495
+000496
+000498
+000499
+000503
+000504
+000506
+000508
+000509
+000510
+000512
+000515
+000517
+000519
+000521
+000524
+000527
+000528
+000530
+000533
+000536
+000541
+000542
+000543
+000545
