+++ tools/cfgs/astyx_models/pointpillar.yaml
