+++ tools/cfgs/dataset_configs/astyx_dataset.yaml
@@ -1,5 +1,5 @@
 DATASET: 'AstyxDataset'
-DATA_PATH: '../data'
+DATA_PATH: '../data/astyx'
 
 POINT_CLOUD_RANGE: [0, -40, -3, 70.4, 40, 1]
